Fix performance issue with many covered areas
Showing that list with a QML Repeater would freeze the app for seconds
if there's several hundred covered areas. And that actually happens for
DWD weather alerts for example.

We can't easily use a ListView instead as this is embedded inside an
already scrollable page. Instead, concatenate inside C++ and and use
a single text element.